Extract common classes into util file
